Refactored header into its own file
