+++ frontend/src/save-formats/GameCube/Components/Basics.js
@@ -1,0 +1,5 @@
+export default class GameCubeHeader {
+  static LITTLE_ENDIAN = false;
+
+  static BLOCK_SIZE = 0x2000;
+}
